+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/Gympass/yoga/compare/@gympass/yoga-icons@0.5.0...@gympass/yoga-icons@0.5.1) (2020-02-18)
+
+**Note:** Version bump only for package @gympass/yoga-icons
+
 # [0.5.0](https://github.com/Gympass/yoga/compare/@gympass/yoga-icons@0.4.0...@gympass/yoga-icons@0.5.0) (2020-02-12)
 
 ### Bug Fixes